Created a path_util.py module and moved the request_file_for_ingress() function there
